+++ resources/lang/de_DE/passwords.php
@@ -1,0 +1,22 @@
+<?php
+
+return [
+
+    /*
+    |--------------------------------------------------------------------------
+    | Password Reset Language Lines
+    |--------------------------------------------------------------------------
+    |
+    | The following language lines are the default lines which match reasons
+    | that are given by the password broker for a password update attempt
+    | has failed, such as for an invalid token or invalid new password.
+    |
+    */
+
+    'reset' => 'Dein Passwort wurde zurückgesetzt',
+    'sent' => 'Es wurde eine Email mit dem Passwortlink veschickt',
+    'throttled' => 'Bitte warten vor einer Wiederholung',
+    'token' => 'Das Passwort-Reset-Token ist ungültig',
+    'user' => "Kein Benutzer mit dieser Emailadresse vorhanden",
+
+];