+++ scripts/installation/CreateLaraProj.sh
@@ -44,22 +44,27 @@ EOS
 show tables;
 EOS
   Wait "Database lrv$PROJ and user $PROJ created"
-  sed -i -e "s/DB_DATABASE=.*/DB_DATABASE=lrv$PROJ/" \
-  -e "s/DB_USERNAME=.*/DB_USERNAME=$PROJ/" \
-  -e "s/DB_PASSWORD=.*/DB_PASSWORD=$PASSW/" .env
-  M_HOST=mail.gmx.net
+  sed -i \
+  -e "s/DB_CONNECTION=.*/DB_CONNECTION=mysql/" \
+  -e "s/^#* *DB_HOST/DB_HOST/" \
+  -e "s/^#* *DB_PORT/DB_PORT/" \
+  -e "s/^#* *DB_DATABASE=.*/DB_DATABASE=lrv$PROJ/" \
+  -e "s/^#* *DB_USERNAME=.*/DB_USERNAME=$PROJ/" \
+  -e "s/^#* *DB_PASSWORD=.*/DB_PASSWORD=$PASSW/" .env
+  M_HOST=smtp.web.de
   M_PORT=587
-  M_USER=laraknife.dev@gmx.net
+  M_USER=laraknife@web.de
   M_PW=Be.Happy4711
   sed -i \
-  -e "s/APP_NAME=.*/APP_NAME=$PROJ"/ \
+  -e "s/APP_NAME=.*/APP_NAME=$PROJ/" \
+  -e "s/APP_LOCALE=en.*/APP_LOCALE=de_DE/" \
   -e "s/MAIL_MAILER=.*/MAIL_MAILER=smtp/" \
   -e "s/MAIL_HOST=.*/MAIL_HOST=$M_HOST/" \
   -e "s/MAIL_PORT=.*/MAIL_PORT=$M_PORT/" \
   -e "s/MAIL_USERNAME=.*/MAIL_USERNAME=$M_USER/" \
   -e "s/MAIL_PASSWORD=.*/MAIL_PASSWORD=$M_PW/" \
   -e "s/MAIL_ENCRYPTION=.*/MAIL_ENCRYPTION=STARTTLS/" \
-  -e "s/MAIL_FROM_ADDRESS=.*/MAIL_FROM_ADDRESS=\"$M_USER\"/" .env
+  -e "s/MAIL_FROM_ADDRESS=.*/MAIL_FROM_ADDRESS=\"\${MAIL_USERNAME}\"/" .env
   sudo sed -i -e "3 i 127.0.0.1 $PROJ.test" /etc/hosts
   grep "$PROJ.test" /etc/hosts
   php artisan migrate
@@ -70,4 +75,5 @@ EOS
   wget "https://public.hamatoma.de/public/IncludeLara.sh" -O $SCRIPT2
   chmod +x $SCRIPT2
   echo "= start $SCRIPT2 git $BASE"
+  php artisan lang:publish
 fi