+++ templates/database/migrations/2023_12_27_071516_modify_users_role.php
@@ -13,6 +13,7 @@ return new class extends Migration
     {
         Schema::table('users', function(Blueprint $table) {
             $table->foreignId('role_id')->nullable()->constrained("roles")->cascadeOnUpdate()->nullOnDelete();
+            $table->string('localization', 8)->default('en_GB');
        });
     }